requirements.txt
fastapi
uvicorn
pypdf
python-docx
requests
pymongo
pinecone-client
google-generativeai
python-dotenv